requirements.txt
Django==5.0.2
django-filter==23.3
djangorestframework==3.14.0
drf-spectacular==0.26.5
psycopg2-binary==2.9.9